+++ gettsim/pre_processing/policy_for_date.py
@@ -6,28 +6,28 @@ from functools import reduce
 import numpy as np
 import yaml
 
-from gettsim.benefits.arbeitsl_geld_2_dag import eink_anr_frei_ab_10_2005
-from gettsim.benefits.arbeitsl_geld_2_dag import eink_anr_frei_bis_10_2005
-from gettsim.benefits.arbeitsl_geld_2_dag import kindersatz_m_ab_2011
-from gettsim.benefits.arbeitsl_geld_2_dag import kindersatz_m_bis_2010
-from gettsim.benefits.arbeitsl_geld_2_dag import regelsatz_m_ab_2011
-from gettsim.benefits.arbeitsl_geld_2_dag import regelsatz_m_bis_2010
-from gettsim.benefits.benefit_checks_dag import kinderzuschlag_temp_vorläufig_ab_2005
-from gettsim.benefits.benefit_checks_dag import kinderzuschlag_temp_vorläufig_bis_2004
-from gettsim.benefits.benefit_checks_dag import regelbedarf_m_vorläufig_ab_2005
-from gettsim.benefits.benefit_checks_dag import regelbedarf_m_vorläufig_bis_2004
-from gettsim.benefits.benefit_checks_dag import wohngeld_basis_hh_vorläufig_ab_2005
-from gettsim.benefits.benefit_checks_dag import wohngeld_basis_hh_vorläufig_bis_2004
-from gettsim.benefits.kinderzuschlag_dag import kinderzuschlag_ab_2005_bis_juni_2019
-from gettsim.benefits.kinderzuschlag_dag import kinderzuschlag_ab_juli_2019
-from gettsim.benefits.kinderzuschlag_dag import kinderzuschlag_eink_regel_ab_2011
-from gettsim.benefits.kinderzuschlag_dag import kinderzuschlag_eink_regel_bis_2010
-from gettsim.benefits.kinderzuschlag_dag import kinderzuschlag_temp_ab_2005
-from gettsim.benefits.kinderzuschlag_dag import kinderzuschlag_temp_bis_2004
-from gettsim.benefits.wohngeld_dag import wohngeld_eink_abzüge_ab_2016
-from gettsim.benefits.wohngeld_dag import wohngeld_eink_abzüge_bis_2015
-from gettsim.benefits.wohngeld_dag import wohngeld_max_miete_ab_2009
-from gettsim.benefits.wohngeld_dag import wohngeld_max_miete_bis_2008
+from gettsim.benefits.arbeitsl_geld_2 import eink_anr_frei_ab_10_2005
+from gettsim.benefits.arbeitsl_geld_2 import eink_anr_frei_bis_10_2005
+from gettsim.benefits.arbeitsl_geld_2 import kindersatz_m_ab_2011
+from gettsim.benefits.arbeitsl_geld_2 import kindersatz_m_bis_2010
+from gettsim.benefits.arbeitsl_geld_2 import regelsatz_m_ab_2011
+from gettsim.benefits.arbeitsl_geld_2 import regelsatz_m_bis_2010
+from gettsim.benefits.benefit_checks import kinderzuschlag_temp_vorläufig_ab_2005
+from gettsim.benefits.benefit_checks import kinderzuschlag_temp_vorläufig_bis_2004
+from gettsim.benefits.benefit_checks import regelbedarf_m_vorläufig_ab_2005
+from gettsim.benefits.benefit_checks import regelbedarf_m_vorläufig_bis_2004
+from gettsim.benefits.benefit_checks import wohngeld_basis_hh_vorläufig_ab_2005
+from gettsim.benefits.benefit_checks import wohngeld_basis_hh_vorläufig_bis_2004
+from gettsim.benefits.kinderzuschlag import kinderzuschlag_ab_2005_bis_juni_2019
+from gettsim.benefits.kinderzuschlag import kinderzuschlag_ab_juli_2019
+from gettsim.benefits.kinderzuschlag import kinderzuschlag_eink_regel_ab_2011
+from gettsim.benefits.kinderzuschlag import kinderzuschlag_eink_regel_bis_2010
+from gettsim.benefits.kinderzuschlag import kinderzuschlag_temp_ab_2005
+from gettsim.benefits.kinderzuschlag import kinderzuschlag_temp_bis_2004
+from gettsim.benefits.wohngeld import wohngeld_eink_abzüge_ab_2016
+from gettsim.benefits.wohngeld import wohngeld_eink_abzüge_bis_2015
+from gettsim.benefits.wohngeld import wohngeld_max_miete_ab_2009
+from gettsim.benefits.wohngeld import wohngeld_max_miete_bis_2008
 from gettsim.config import ROOT_DIR
 from gettsim.pre_processing.piecewise_functions import get_piecewise_parameters
 from gettsim.pre_processing.policy_completion_funcs import add_progressionsfaktor
